(#2) ⚒️ Refactor: Layout 폴더 생성
