+++ src/app/layout.tsx
@@ -2,8 +2,8 @@ import type { Metadata } from 'next'
 import localFont from 'next/font/local'
 import './globals.css'
 
-import Header from '@/components/Header'
-import Footer from '@/components/Footer'
+import Header from '@/components/Layout/Header'
+import Footer from '@/components/Layout/Footer'
 
 const pretendard = localFont({
   src: [
