+++ src/app/page.tsx
@@ -1,7 +1,7 @@
 export default function Home() {
   return (
     <div className="text-primaryT-90">
-      <p className="bold-36">HOME</p>
+      <p className="bold-36 text-black">HOME</p>
     </div>
   )
 }
