+++ src/components/Layout/Footer.tsx
@@ -1,3 +1,3 @@
 export default function Footer() {
-  return <footer>FOOTER</footer>;
+  return <div>FOOTER</div>;
 }
